mod: mv prover option from utils to prover
