+++ crates/utils/src/lib.rs
@@ -2,7 +2,6 @@ pub mod initializer;
 pub mod logging;
 pub mod network;
 pub mod options;
-pub mod prover_options;
 pub mod prover_types;
 pub mod rpc;
 pub mod unique_heap;
